refactor: split runner xctestrun module
